1. Move `LegendItem` to a separate file. 2. Fix typo.
